Unusual basketball scorecard - STACKS (EASY)
This is the basic first upload of the code. This is an Easy question:
In this question, the unusual part is that  we are given a list of strings operations, where operations[i] is the 'i'th operation you must apply to the record and is one of the following:

An integer x.
Record a new score of x.
'+'.
Record a new score that is the sum of the previous two scores.
'D'.
Record a new score that is the double of the previous score.
'C'.
Invalidate the previous score, removing it from the record.
Return the sum of all the scores on the record after applying all the operations.

The test cases are generated such that the answer and all intermediate calculations fit in a 32-bit integer and that all operations are valid.